requirements.txt
aiohttp==3.6.2
asgiref @ file:///home/conda/feedstock_root/build_artifacts/asgiref_1592528583041/work
async-timeout==3.0.1
attrs==20.2.0
beautifulsoup4 @ file:///D:/bld/beautifulsoup4_1589761031242/work
brotlipy==0.7.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
defusedxml==0.6.0
diff-match-patch==20200713
Django==2.2.1
django-crispy-forms==1.9.2
django-formtools==2.2
django-import-export==2.3.0
django-polls==0.1
django-ranged-response==0.2.0
django-reversion==3.0.7
django-simple-captcha==0.5.12
et-xmlfile==1.0.1
future==0.18.2
httplib2==0.9.2
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
idna-ssl==1.1.0
jdcal==1.4.1
MarkupPy==1.14
multidict==4.7.6
mysqlclient==1.3.14
odfpy==1.4.1
olefile==0.46
openpyxl==3.0.4
Pillow==7.1.2
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
PyMySQL==0.9.3
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1594392929924/work
PySocks==1.7.1
pytz==2020.1
pywin32==227
PyYAML==5.3.1
reportlab==3.5.44
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
six==1.15.0
soupsieve @ file:///D:/bld/soupsieve_1589778616374/work
sqlparse==0.3.1
tablib==2.0.0
tqdm==4.48.2
typing-extensions==3.7.4.3
urllib3==1.25.9
win-inet-pton==1.1.0
wincertstore==0.2
xadmin @ file:///F:/django/python%E8%B5%84%E6%96%99/xadmin-master.zip
xlrd==1.2.0
xlwt==1.3.0
yarl==1.5.1
